// src/generated/model/studentVisaResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface StudentVisaResponse {
  id?: string;
  studentId?: string;
  visaNumber?: string;
  passportNumber?: string;
  countryOfIssue?: string;
  issueDate?: string;
  expiryDate?: string;
  sponsor?: string;
  status?: string;
  remarks?: string;
  createdBy?: string;
  updatedBy?: string;
  createdAt?: string;
  updatedAt?: string;
}